Change Calculator functions

+++ UtilityKnife/UtilityKnife/Calculator.cs
@@ -13,12 +13,12 @@ namespace UtilityKnife
             return a + b;
         }
 
-        public static double Subtract(int a, int b)
+        public static double Subtract(double a, double b)
         {
             return a - b;
         }
 
-        public static int Multiply(int a, int b)
+        public static double Multiply(double a, double b)
         {
             return a * b;
         }
